- interface.c: add "interface" config for unguessable interface type/speed
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@4679 06827809-a52a-0410-b366-d66718629ded

+++ agent/mibgroup/mibII/interfaces.h
@@ -61,7 +61,7 @@ struct ifnet {
 	char    if_hwaddr [6];		/* ethernet address */
 	int	if_type;		/* interface type: 1=generic,
 					   28=slip, ether=6, loopback=24 */
-	int	if_speed;		/* interface speed: in bits/sec */
+	u_long	if_speed;		/* interface speed: in bits/sec */
 
 	struct sockaddr if_addr;	/* interface's address */
 	struct sockaddr ifu_broadaddr;	/* broadcast address */
